R-Week-4 Assignment
Your assignment should be published to RPubs.com. Place your script in a
public repository on GitHub and submit
a link to the URL here.
You may work with up to three partners of your own choosing on this
assignment. Group work should be more
impressive.

Exploratory Data Analysis in R. Choose an interesting dataset and use R
graphics to describe the data. You may
use base R graphics, or a graphics package of your choice. You should
include at least one example of each of the
following:
• histogram
• boxplot
• scatterplot
Do the graphics provide insight into any relationships in the data?
You may use R code that already exists in a book or a web site as a
starting point for your work. You need to cite
your source, and you need to clearly delineate what you were given to
start with, and what is your own original
analysis, so that your grade can be based on your own work!
This is an open-ended assignment. Full credit will be given for any
reasonable effort, but as with most technical
learning work, the more sustained effort over time that you invest, the
more you will get out of it.
If you don’t know where to start, there is a list of over 700 R datasets
here:
http://vincentarelbundock.github.io/Rdatasets/. You can of course also
load data from another source, such as a
.CSV file. You’ll want to pick a dataset where it makes sense to use
histograms, boxplots, and scatterplots.